refactor: decompose AppSidebar and EditorToolbar
Extract focused sub-components from two 1000+ line files:

AppSidebar (1028 → ~330 lines):
- sidebar/InitiativeSection.tsx
- sidebar/TagBrowser.tsx
- sidebar/SidebarUserFooter.tsx

EditorToolbar (1080 → ~530 lines):
- editor/toolbar/editorToolbarUtils.ts
- editor/toolbar/EditorMobileOverflowMenu.tsx
- editor/toolbar/EditorTableMenu.tsx
- editor/toolbar/EditorInsertMenu.tsx

No behavioral changes. Closes #172
